Fix JSON writing and checking.

+++ verify_single.sh
@@ -8,10 +8,14 @@ for i in `find expected/ -type f`; do
 	testfile=${i/expected/test.latest}
 	sed -i -e 's/Exception(test timed out after \([^ ]*\) milliseconds): [^"]*/TimeoutException after \1 ms/g' $testfile
 	sed -i -e 's/StackOverflowError(): [^"]*/StackOverflowError()/g' $testfile
-	if [[ "$i" == expected/run*.err ]]; then
-		# pretty print as json before diffing
+	if [[ "$i" == expected/run*.err ]] && [[ -s "$testfile" ]]; then
+		# pretty print as json before diffing (if size > 0)
 		cat $testfile | python -m json.tool > ${testfile}.new
-		mv ${testfile}.new $testfile
+		if [[ $? -ne 0 ]]; then
+			echo -e "Above JSON is broken" >> $testfile
+		else
+			mv ${testfile}.new $testfile
+		fi
 	fi
 	diff -b -u -I '^make' -I '^Makefile:' $i $testfile
 	ec=$?
